#trivial remove redundant types for request / method / action (#1248)
* reorg message

* deprecate Action

* replace with Method

* cleanup

* as Method
