+++ packages/wallet-sdk/src/sign/scw/SCWSigner.ts
@@ -3,11 +3,9 @@ import { SCWKeyManager } from './SCWKeyManager';
 import { SCWStateManager } from './SCWStateManager';
 import { PopUpCommunicator } from ':core/communicator/PopUpCommunicator';
 import { standardErrors } from ':core/error';
-import { createMessage } from ':core/message';
-import { Action, SupportedEthereumMethods, SwitchEthereumChainAction } from ':core/message/Action';
-import { RPCRequestMessage, RPCResponseMessage } from ':core/message/RPCMessage';
-import { RPCResponse } from ':core/message/RPCResponse';
+import { createMessage, RPCRequestMessage, RPCResponse, RPCResponseMessage } from ':core/message';
 import { AppMetadata, RequestArguments } from ':core/provider/interface';
+import { Method } from ':core/provider/method';
 import { AddressString } from ':core/type';
 import { ensureIntNumber } from ':core/type/util';
 import {
@@ -16,6 +14,12 @@ import {
   exportKeyToHexString,
   importKeyFromHexString,
 } from ':util/cipher';
+
+type SwitchEthereumChainParam = [
+  {
+    chainId: `0x${string}`; // Hex chain id
+  },
+];
 
 export class SCWSigner implements Signer {
   private readonly metadata: AppMetadata;
@@ -47,7 +51,7 @@ export class SCWSigner implements Signer {
 
     const handshakeMessage = await this.createRequestMessage({
       handshake: {
-        method: SupportedEthereumMethods.EthRequestAccounts,
+        method: 'eth_requestAccounts',
         params: this.metadata,
       },
     });
@@ -61,7 +65,7 @@ export class SCWSigner implements Signer {
     await this.keyManager.setPeerPublicKey(peerPublicKey);
 
     const decrypted = await this.decryptResponseMessage<AddressString[]>(response);
-    this.updateInternalState({ method: SupportedEthereumMethods.EthRequestAccounts }, decrypted);
+    this.updateInternalState({ method: 'eth_requestAccounts' }, decrypted);
 
     const result = decrypted.result;
     if ('error' in result) throw result.error;
@@ -93,9 +97,9 @@ export class SCWSigner implements Signer {
   }
 
   private tryLocalHandling<T>(request: RequestArguments): T | undefined {
-    switch (request.method) {
-      case SupportedEthereumMethods.WalletSwitchEthereumChain: {
-        const params = request.params as SwitchEthereumChainAction['params'];
+    switch (request.method as Method) {
+      case 'wallet_switchEthereumChain': {
+        const params = request.params as SwitchEthereumChainParam;
         if (!params || !params[0]?.chainId) {
           throw standardErrors.rpc.invalidParams();
         }
@@ -105,7 +109,7 @@ export class SCWSigner implements Signer {
         // https://eips.ethereum.org/EIPS/eip-3326#wallet_switchethereumchain
         return switched ? (null as T) : undefined;
       }
-      case SupportedEthereumMethods.WalletGetCapabilities: {
+      case 'wallet_getCapabilities': {
         const walletCapabilities = this.stateManager.walletCapabilities;
         if (!walletCapabilities) {
           // This should never be the case for scw connections as capabilities are set during handshake
@@ -130,7 +134,7 @@ export class SCWSigner implements Signer {
 
     const encrypted = await encryptContent(
       {
-        action: request as Action,
+        action: request,
         chainId: this.stateManager.activeChain.id,
       },
       sharedSecret
@@ -184,18 +188,18 @@ export class SCWSigner implements Signer {
     const result = response.result;
     if ('error' in result) return;
 
-    switch (request.method) {
-      case SupportedEthereumMethods.EthRequestAccounts: {
+    switch (request.method as Method) {
+      case 'eth_requestAccounts': {
         const accounts = result.value as AddressString[];
         this.stateManager.updateAccounts(accounts);
         break;
       }
-      case SupportedEthereumMethods.WalletSwitchEthereumChain: {
+      case 'wallet_switchEthereumChain': {
         // "return null if the request was successful"
         // https://eips.ethereum.org/EIPS/eip-3326#wallet_switchethereumchain
         if (result.value !== null) return;
 
-        const params = request.params as SwitchEthereumChainAction['params'];
+        const params = request.params as SwitchEthereumChainParam;
         const chainId = ensureIntNumber(params[0].chainId);
         this.stateManager.switchChain(chainId);
         break;
